+++ src/map_builders/waveform_collapse/solver.rs
@@ -166,10 +166,10 @@ impl Solver {
             // There is nothing nearby, so we can have anything!
             let new_chunk_idx = (rng.roll_dice(1, self.constraints.len() as i32) - 1) as usize;
             self.chunks[chunk_index] = Some(new_chunk_idx);
-            let left_x = chunk_x as i32 * self.chunk_size as i32;
-            let right_x = (chunk_x + 1) as i32 * self.chunk_size as i32;
-            let top_y = chunk_y as i32 * self.chunk_size as i32;
-            let bottom_y = (chunk_y + 1) as i32 * self.chunk_size as i32;
+            let left_x = chunk_x as i32 * self.chunk_size;
+            let right_x = (chunk_x + 1) as i32 * self.chunk_size;
+            let top_y = chunk_y as i32 * self.chunk_size;
+            let bottom_y = (chunk_y + 1) as i32 * self.chunk_size;
 
             let mut i: usize = 0;
             for y in top_y..bottom_y {
@@ -214,10 +214,10 @@ impl Solver {
                 };
 
                 self.chunks[chunk_index] = Some(possible_options[new_chunk_idx as usize]);
-                let left_x = chunk_x as i32 * self.chunk_size as i32;
-                let right_x = (chunk_x + 1) as i32 * self.chunk_size as i32;
-                let top_y = chunk_y as i32 * self.chunk_size as i32;
-                let bottom_y = (chunk_y + 1) as i32 * self.chunk_size as i32;
+                let left_x = chunk_x as i32 * self.chunk_size;
+                let right_x = (chunk_x + 1) as i32 * self.chunk_size;
+                let top_y = chunk_y as i32 * self.chunk_size;
+                let bottom_y = (chunk_y + 1) as i32 * self.chunk_size;
 
                 let mut i: usize = 0;
                 for y in top_y..bottom_y {
